CR restructure
